I was running into some performance issues while using `VictoryVoronoiContainer` so I did some investigation, and found that throttling VoronoiHelpers.onMouseMove didn't actually throttle as expected. So it looks like when throttling a function it will return the value that was last returned by the function, so the container was returning duplicate mutations in `defaultEvents.onMouseMove` and causing unnecessary rerenders.

Here are some Chrome timelines to show the perf impact

**Before**
Throttle 16ms (default)
![screen shot 2017-03-16 at 3 15 11 pm](https://cloud.githubusercontent.com/assets/8966794/24022334/18252696-0a63-11e7-8386-c1e5cf1fa6c9.png)

Throttle 3000ms
![screen shot 2017-03-16 at 3 16 22 pm](https://cloud.githubusercontent.com/assets/8966794/24022340/23c0242e-0a63-11e7-83c1-64a7ab14fc0c.png)

**After**
Throttle 16ms
![screen shot 2017-03-16 at 3 12 02 pm](https://cloud.githubusercontent.com/assets/8966794/24022350/320802e0-0a63-11e7-9d2b-bbf61c089486.png)

Throttle 32ms
![screen shot 2017-03-16 at 3 12 48 pm](https://cloud.githubusercontent.com/assets/8966794/24022357/3d60fea8-0a63-11e7-86a8-0abe5220fc23.png)

Throttle 3000ms
![screen shot 2017-03-16 at 3 13 41 pm](https://cloud.githubusercontent.com/assets/8966794/24022365/459b86ec-0a63-11e7-9acc-3e749fab7a56.png)

Also increased the throttling to 32ms, I didn't notice any impact on the responsiveness of the tooltips.