@nfcampos 2017-03-18
I noticed that you'd most likely get performance benefits out of not persisting onMouseMove events here https://github.com/FormidableLabs/victory-chart/blob/master/src/components/containers/victory-voronoi-container.js#L37,
which would be made possible by adding `trailing: false` to https://github.com/FormidableLabs/victory-chart/blob/master/src/components/containers/voronoi-helpers.js#L166.
What I'm not sure is whether it'd make the experience worse in any way, because it does mean that, in the worst case scenario, the last 31ms of updates to the mouse position are ignored.